+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.7.2] - 2019-06-27
+
 ### Fixed
 
 - Avoid passing classes that are not typography tokens to the `font` prop.